Refactor container test fixtures for PSR-4
