+++ composer.json
@@ -47,9 +47,6 @@
             "GustavPHP\\Demo\\": "demo/",
             "GustavPHP\\Tests\\": "tests/"
         },
-        "exclude-from-classmap": [
-            "tests/Unit/Service/ContainerTest.php"
-        ],
         "files": [
             "tests/Integration/helpers.php"
         ]
