+++ tests/Fixtures/Service/ContainerTestCircularA.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+class ContainerTestCircularA
+{
+    public function __construct(public ContainerTestCircularB $b)
+    {
+    }
+}
