+++ tests/Fixtures/Service/ContainerTestCircularB.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+class ContainerTestCircularB
+{
+    public function __construct(public ContainerTestCircularA $a)
+    {
+    }
+}
