+++ tests/Fixtures/Service/ContainerTestCircularController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+use GustavPHP\Gustav\Controller\Base;
+
+class ContainerTestCircularController extends Base
+{
+    public function __construct(public ContainerTestCircularA $a)
+    {
+    }
+}
