+++ tests/Fixtures/Service/ContainerTestDefinitionController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+use GustavPHP\Gustav\Controller\Base;
+
+class ContainerTestDefinitionController extends Base
+{
+    public function __construct(public ContainerTestProvidedService $provided)
+    {
+    }
+}
