+++ tests/Fixtures/Service/ContainerTestEmptyController.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+use GustavPHP\Gustav\Controller\Base;
+
+class ContainerTestEmptyController extends Base
+{
+}
