+++ tests/Fixtures/Service/ContainerTestNestedDependency.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+class ContainerTestNestedDependency
+{
+    public function __construct(public ContainerTestPlainDependency $plain)
+    {
+    }
+}
