+++ tests/Fixtures/Service/ContainerTestPlainDependency.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+class ContainerTestPlainDependency
+{
+}
