+++ tests/Fixtures/Service/ContainerTestProvidedService.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+class ContainerTestProvidedService
+{
+    public function __construct(public string $value)
+    {
+    }
+}
