+++ tests/Fixtures/Service/ContainerTestUnresolvableController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace GustavPHP\Tests\Fixtures\Service;
+
+use GustavPHP\Gustav\Controller\Base;
+
+class ContainerTestUnresolvableController extends Base
+{
+    public function __construct(public string $value)
+    {
+    }
+}
