+++ tests/Unit/Service/ContainerTest.php
@@ -1,7 +1,7 @@
 <?php
 
-use GustavPHP\Gustav\Controller\Base;
 use GustavPHP\Gustav\Service\Container;
+use GustavPHP\Tests\Fixtures\Service\{ContainerTestAutowiredController, ContainerTestCircularController, ContainerTestDefinitionController, ContainerTestEmptyController, ContainerTestNestedDependency, ContainerTestPlainDependency, ContainerTestProvidedService, ContainerTestUnresolvableController};
 
 it('requires the container to be built before resolving controllers', function () {
     $container = new Container();
@@ -71,87 +71,3 @@ it('throws when a constructor parameter cannot be resolved', function () {
 
     $container->make(ContainerTestUnresolvableController::class);
 })->throws(InvalidArgumentException::class);
-
-if (!class_exists(ContainerTestPlainDependency::class)) {
-    class ContainerTestPlainDependency
-    {
-    }
-}
-
-if (!class_exists(ContainerTestNestedDependency::class)) {
-    class ContainerTestNestedDependency
-    {
-        public function __construct(public ContainerTestPlainDependency $plain)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestAutowiredController::class)) {
-    class ContainerTestAutowiredController extends Base
-    {
-        public function __construct(public ContainerTestNestedDependency $dependency)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestProvidedService::class)) {
-    class ContainerTestProvidedService
-    {
-        public function __construct(public string $value)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestDefinitionController::class)) {
-    class ContainerTestDefinitionController extends Base
-    {
-        public function __construct(public ContainerTestProvidedService $provided)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestEmptyController::class)) {
-    class ContainerTestEmptyController extends Base
-    {
-    }
-}
-
-if (!class_exists(ContainerTestCircularA::class)) {
-    class ContainerTestCircularA
-    {
-        public function __construct(public ContainerTestCircularB $b)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestCircularB::class)) {
-    class ContainerTestCircularB
-    {
-        public function __construct(public ContainerTestCircularA $a)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestCircularController::class)) {
-    class ContainerTestCircularController extends Base
-    {
-        public function __construct(public ContainerTestCircularA $a)
-        {
-        }
-    }
-}
-
-if (!class_exists(ContainerTestUnresolvableController::class)) {
-    class ContainerTestUnresolvableController extends Base
-    {
-        public function __construct(public string $value)
-        {
-        }
-    }
-}
